Refactor: Organize chat-room API into internal modules
