+++ api/chat-rooms/_helpers.ts
