+++ public/version.json
@@ -1,8 +1,8 @@
 {
   "version": "10.3",
-  "buildNumber": "6a044cc",
-  "commitSha": "6a044cc1a329aded1472bc2eb9664a049e22d9bb",
-  "buildTime": "2025-12-06T13:14:41.197Z",
+  "buildNumber": "7c1d246",
+  "commitSha": "7c1d24650a278bbf5d50df1a371fc837b7e739b4",
+  "buildTime": "2025-12-06T16:11:20.569Z",
   "majorVersion": 10,
   "minorVersion": 3
 }
